pyproject.toml
[project]
name = "paper-writer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "markdownify>=1.1.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "typing>=3.10.0.0",
]
